refactor: Splits the documentspage into components - Adds AdminContentList component. - Adds AdminHeader component. - Adds AdminToolbar component. - Adds Count component. Issue #282
